Add hooks to emulator and rework structure
