+++ photon/libphoton/photon/photon-hooks.hh
@@ -1,0 +1,25 @@
+#pragma once
+#include <vector>
+#include <functional>
+#include <stdint.h>
+
+namespace photon {
+
+template<class... Args>
+struct PhotonHook {
+    std::vector<std::function<void(Args...)>> callbacks;
+
+    void operator()(Args... args) {
+        for (auto& f : callbacks) {
+            f(args...);
+        }
+    }
+
+};
+
+struct PhotonHooks {
+    PhotonHook<size_t, uintptr_t> on_mem_read;
+    // TODO: Declare hooks here
+};
+
+}
